+++ src/Servidor/Modelo/NormalWeapon.h
@@ -14,7 +14,7 @@ class Npc;
 class NormalWeapon : public Weapon {
     public:
         NormalWeapon(std::string name, int minDamage,int maxDamage, int goldCost);
-        int attack(Character *enemy,int strength,int level,int &mana,Position &holderPos)  override;
+        int attack(Character *enemy,int strength,int level,float &mana,Position &holderPos)  override;
         ~NormalWeapon() override;
 };
