+++ src/Servidor/Modelo/RangeWeapon.cpp
@@ -7,7 +7,7 @@ RangeWeapon::RangeWeapon(std::string name, int minDamage, int maxDamage, int gol
     :Weapon(name, minDamage,maxDamage, goldCost) {
 }
 
-int RangeWeapon::attack(Character *enemy,int strength,int level,int &mana,Position &holderPos) {
+int RangeWeapon::attack(Character *enemy,int strength,int level,float &mana,Position &holderPos) {
     int damage = calculateDamage(strength);
     return enemy->receiveDamage(level, damage);
 }
